Created Viewprofile.ItemTemplate2, Viewprofile, image.jpg and Viewprofile.ItemTemplate1; Edited customer; Deleted customer.ItemTemplate2 and customer.ItemTemplate1
