+++ .anvil_editor.yaml
@@ -3,10 +3,11 @@ unique_ids:
     Form1: 27OJC6EWBTRQ34P5OQF2I5IMMQAQK7SC
     LOGIN: '1700197874946586108155809.6506'
     SIGNUP: '1700198434288731474488224.3363'
+    Viewprofile: '1700199596763230077997025.92548'
+    Viewprofile.ItemTemplate1: '1700558077194623003567799.6544'
+    Viewprofile.ItemTemplate2: '1700558120994607838817366.404'
     admin: '1700199610247838843848268.7214'
-    customer: '1700199596763230077997025.92548'
-    customer.ItemTemplate1: '1700558077194623003567799.6544'
-    customer.ItemTemplate2: '1700558120994607838817366.404'
+    customer: '1700650184946263831557477.6373'
   modules: {}
   server_modules:
     ServerModule1: '1700198830978493420714307.2622'
@@ -40,4 +41,5 @@ unique_ids:
     digital-wallet-smartphone-icon-made-600nw-2314309781.webp: '1699969269890961734173005.2742'
     bus-driver_8792037.png: '1700147718670700965961084.8494'
     new1.jpg: '1700051745425827756062277.6134'
+    image.jpg: '1700650300027448163955744.21277'
     travel_7969556.png: '1700146069263280886013499.4022'
